+++ src/parser/element_section.rs
@@ -24,7 +24,7 @@ impl Parser {
         println!("  # Parsing element section");
         let init_offset = self.get_current_offset();
         let entries = self.read_vu32_times(Parser::read_element);
-        assert!(self.get_read_len(init_offset) == payload_len as u64);
+        assert!(self.get_read_len(init_offset) == payload_len);
         println!("{:?}", entries);
         println!("  + Parsing element section done");
         // return entries;
