+++ src/parser/function_section.rs
@@ -9,7 +9,7 @@ impl Parser {
         println!("  # Parsing function section");
         let init_offset = self.get_current_offset();
         let ids = self.read_vu32_times(Parser::read_fn_id);
-        assert!(self.get_read_len(init_offset) == payload_len as u64);
+        assert!(self.get_read_len(init_offset) == payload_len);
         println!("{:?}", ids);
         println!("  + Parsing function section done");
         // return types
