+++ src/parser/table_section.rs
@@ -16,7 +16,7 @@ impl Parser {
         println!("  # Parsing table section");
         let init_offset = self.get_current_offset();
         let entries = self.read_vu32_times(Parser::read_table_type);
-        assert!(self.get_read_len(init_offset) == payload_len as u64);
+        assert!(self.get_read_len(init_offset) == payload_len);
         println!("{:?}", entries);
         println!("  + Parsing table section done");
         //return entries
